Bugfix issue #22
Not sure when this error appeared, but its presence in an_NPLprint, an_USCprint & an_NEDprint suggests that it's old and have created a cascade of problems that were unintentionally fixed with the deletion routine in analysis.m

Also, I guess this would only throw an exception on macro 910 (vfloat + sweeps on LAP1). This bug wouldn't have been detected if I'd defined some arbitrary array a row array instead of a column array, and would have been much much harder to locate.

Also, an unrelated bug was found in  analysis.m . Let's hope Eriks's LBL code is up to the task for NED.TAB files