Wheel-off loading implementation (at last!) + NEL qv
The accuracy of the WOL dataset limits the accuracy of the WOL
detection. We don't have exact durations of WOL, so we set the max
duration to what MAG typically sees from the when the disturbances of
the circuitry stops. (198 seconds)

We then flag all measurements in the CALIB file if any point is within
this window

NEL got some updated quality values. Unfortunately I didn’t have time
to test all of these