untested: MIP_XCAL interval filter
Finally recieved the intervals for MIP xcal, which is to be filtered out in the dataset.
unverified, untested, unbroken.
Should work as long as t_et is as long as the data array

Also, added some logic where files that will be empty will not be created at all (and thus removed) instead of being printed and then deleted.