Fix for issue #51. Also NED bugfix for Macro 710/910
710 interpolation routine (for calibrating USC>NED) inherited LAP2 time
stamps (which is sparser than LAP1). which caused the interpolation
and associated QVs to give NAN values (which were correctly changed to
-1e9) and the quality value to give NAN

The second bug is from an erroneous NEDFlag (i.e. probe number),
stemming from a Matlab bug (with no error message). when LAP1 LAP2 in
e-field mode, and LAP1 is shadowed.

Bugs confined to NED files in ESC2- EXT3